Change Type Data in DB, Create Employee Function Done

+++ database/migrations/2019_06_12_095938_create_balances_table.php
@@ -16,7 +16,7 @@ class CreateBalancesTable extends Migration
         Schema::create('balances', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->bigInteger('user_id')->unsigned();
-            $table->float('nominal')->default(0.0);
+            $table->integer('nominal')->default(0);
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users');
